Separate items from CompoundItemItemProperties.cs
