+++ Fuyu.Backend.BSG/ItemTemplates/EEquipmentSlot.cs
@@ -1,0 +1,20 @@
+﻿namespace Fuyu.Backend.BSG.ItemTemplates;
+
+public enum EEquipmentSlot
+{
+    FirstPrimaryWeapon,
+    SecondPrimaryWeapon,
+    Holster,
+    Scabbard,
+    Backpack,
+    SecuredContainer,
+    TacticalVest,
+    ArmorVest,
+    Pockets,
+    Eyewear,
+    FaceCover,
+    Headwear,
+    Earpiece,
+    Dogtag,
+    ArmBand
+}
